requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
python-multipart==0.0.9
python-dotenv==1.0.1
sqlalchemy==2.0.32
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
openai==1.51.0
pillow==10.4.0
pandas==2.2.2
matplotlib==3.9.0
yfinance==0.2.43
# Optional MCP integration
mcp==1.2.0
